Separate wrapper building functions to distinct files for parallel build
Not all wrapper building functions are separated.
