+++ tests/test_mesh.py
@@ -147,18 +147,18 @@ class StaticMeshTC(unittest.TestCase):
         # viewer.
         np.testing.assert_almost_equal(
             mh.fccnd,
-            [[-0.3333333,  0.6666667,  0.       ],
-             [ 0.       ,  0.6666667,  0.3333333],
-             [-0.3333333,  0.6666667,  0.3333333],
-             [-0.3333333,  1.       ,  0.3333333]])
+            [[-0.3333333,  0.6666667,  0.       ],  # noqa
+             [ 0.       ,  0.6666667,  0.3333333],  # noqa
+             [-0.3333333,  0.6666667,  0.3333333],  # noqa
+             [-0.3333333,  1.       ,  0.3333333]])  # noqa
         np.testing.assert_almost_equal(
             mh.fcnml,
-            [[ 0.       ,  0.       , -1.       ],
-             [ 1.       ,  0.       ,  0.       ],
-             [-0.5773503, -0.5773503,  0.5773503],
-             [ 0.       ,  1.       ,  0.       ]])
+            [[ 0.       ,  0.       , -1.       ],  # noqa
+             [ 1.       ,  0.       ,  0.       ],  # noqa
+             [-0.5773503, -0.5773503,  0.5773503],  # noqa
+             [ 0.       ,  1.       ,  0.       ]])  # noqa
         np.testing.assert_almost_equal(
-            mh.fcara, [0.5      , 0.5      , 0.8660254, 0.5      ])
+            mh.fcara, [0.5      , 0.5      , 0.8660254, 0.5      ])  # noqa
         np.testing.assert_almost_equal(
             mh.clcnd, [[-0.25,  0.75,  0.25]])
         np.testing.assert_almost_equal(
